Renamed columns and relationships in the code
There were quite a few places that had generic 'user_id' columns to make coding
up the relations quicker to start with.  But now renamed to more readable names
like 'staff_id', 'student_id'.  Relations on models updated to match - so for
instance instead of $assessment->user() it's $assessment->staff() which is a little
easier to grok.