Added ability to remove data from the system that isn't in the wlm
The wlm importer can now run in two ways - a straight import of
the data into the local db, and a 'sync' of the data which removes
any courses, students or staff who aren't in the wlm but exist
locally.

This is needed so we don't end up with years of old data in the
system which would eventually slow down all of the db queries.

Now to decide if the sync is the default, or the additive 'run'
is the default...