Some docblocks and a tiny refactor
Refactor was just in the assessment model where student feedback is
recorded.  Before it was creating a new feedback if one didn't already
exist, filling the fields and saving it.  But it also filled the fields
and saved it if it was a duplicate - so just made it return before
saving rather than re-saving.  As it were... ;-)

In theory it should never happen anyway, but it was a bit silly to leave
it in.