Removed all the model event listeners
Had a load of model event listeners which would delete related records
when a model was being deleted.  These only existed as sqlite defaults
to ignoring foreign key information on tables.  Found a way of making
laravel enable support in sqlite so switched to that - all the
model event code could now be removed.

Sqlite over-ride is in AppServiceProvider if needed elsewhere.