Work on active/inactive courses
Courses in the WLM can be marked as 'current' or not - this means they
shouldn't show up locally in terms of assessments, calendars etc.

So a bit of work on adding a flag to the local Course model to indicate
that which is set when importing from the WLM.  That flag is honoured
by the various bits of code which show assessments to students & staff.